refactor: rename sample directory to examples
