+++ examples/source-example/src/main/java/com/vance/samples/MyAdapter.java
@@ -1,0 +1,25 @@
+package com.vance.samples;
+
+import com.linkall.vance.core.Adapter1;
+import io.cloudevents.CloudEvent;
+import io.cloudevents.core.builder.CloudEventBuilder;
+
+import java.net.URI;
+import java.time.OffsetDateTime;
+import java.util.UUID;
+
+public class MyAdapter implements Adapter1<String> {
+    private static final CloudEventBuilder template = CloudEventBuilder.v1();
+    @Override
+    public CloudEvent adapt(String data) {
+        template.withId(UUID.randomUUID().toString());
+        URI uri = URI.create("vance-http-source");
+        template.withSource(uri);
+        template.withType("http");
+        template.withDataContentType("application/json");
+        template.withTime(OffsetDateTime.now());
+        template.withData(data.getBytes());
+
+        return template.build();
+    }
+}
